% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crespO2.R
\name{crespO2}
\alias{crespO2}
\title{C_resp value to oxygen consumption}
\usage{
crespO2(C_resp, C, k)
}
\arguments{
\item{C_resp}{(class numeric) A C_resp value - the proportion of metabolic carbon in a fish's blood.
Must be between 0 and 1.}

\item{C}{(class numeric) The upper bound of C_resp.}

\item{k}{(class numeric) The decay constant of the relationship between C_resp and oxygen consumption}
}
\value{
(class numeric) Oxygen consumption in mg O2/kg/h.
}
\description{
Calculates mass-specific oxygen consumption (mg O2/kg/h) from a C_resp value.
}
